requirements.txt
fastapi
uvicorn
gradio-client
Pillow
retry
cachetools
python-dotenv
